requirements.txt
streamlit
pandas
numpy
requests
openai
langchain
chromadb
sentence-transformers
# Add other packages your project actually uses
google-genai
voyageai
faiss-cpu
